Izmenjena skripta
Refaktorisan emailService za slanje notifikacija
