+++ backend/src/main/resources/application.properties
@@ -2,7 +2,7 @@
 spring.datasource.url=jdbc:mysql://localhost:3306/Ticketist?useSSL=false&createDatabaseIfNotExist=true
 spring.datasource.username=root
 spring.datasource.password=admin
-spring.datasource.initialization-mode=always
+#spring.datasource.initialization-mode=always
 
 spring.jpa.show-sql=true
 spring.jpa.hibernate.format_sql=true
