+++ backend/src/main/resources/data.sql
@@ -1,56 +1,57 @@
-TRUNCATE TABLE ticketist.venues;
-TRUNCATE TABLE ticketist.events;
-TRUNCATE TABLE ticketist.tickets;
+TRUNCATE TABLE venues;
+TRUNCATE TABLE events;
+TRUNCATE TABLE tickets;
+TRUNCATE TABLE users;
 
 INSERT INTO venues (id, is_active, name, street, city, latitude, longitude)
 VALUES
-	(1, true, "Spens", "ulica", "grad", 1, 1),
-    (2, true, "Novi Sad Fair", "ulica", "grad", 1, 1),
-    (3, true, "Startit Centar", "ulica", "grad", 1, 1);
+	(1, true, 'Spens', 'Sutjeska 2', 'Novi Sad', 45.2470, 19.8453),
+    (2, true, 'Novi Sad Fair', 'Hajduk Veljkova 11', 'Novi Sad', 45.267136, 19.833549),
+    (3, true, 'Startit Centar', 'Miroslava Antića 2', 'Novi Sad', 45.267136, 19.833549);
 
 INSERT INTO events
-	(id, name, description, start_date, end_date, reservation_deadline, reservation_limit, venue_id)
+	(id, name, description, category, start_date, end_date, reservation_deadline, reservation_limit, venue_id)
 VALUES
-	(4, "Event1", "Opis eventa 1", CURDATE(), CURDATE(), '2019-10-05', 3, 1),
-    (5, "Event2", "Opis eventa 2", '2019-12-12', '2019-12-20', '2019-10-05', 3, 1),
-    (6, "Event3", "Opis eventa 3", CURDATE(), CURDATE(), '2019-10-05', 3, 2),
-    (7, "Event4", "Opis eventa 4", '2019-12-12', '2019-12-20', '2019-10-05', 3, 2),
+	(1, 'Danasnji event', 'Opis eventa koji se odrzava danas na Spensu', 'SPORTS', CURDATE(), CURDATE(), CURDATE(), 3, 1),
+    (2, 'Koncert Adila', 'Koncert Adila', 'ENTERTAINMENT', '2019-11-12', '2019-11-12', '2019-11-01', 3, 1),          -- potrebno dodati sate odrzavanja
+    (3, 'Sajam knjiga', 'Novosadski sajam knjiga', 'CULTURAL', '2020-03-05', '2020-03-11', '2020-03-01', 3, 2),
+    (4, 'Poljoprivredni sajam', 'Medjunarodni sajam poljoprivrede', 'CULTURAL', '2020-05-11', '2020-05-17', '2020-05-01', 3, 2),
 
-    (8, "Spens event 1", "Opis spens eventa 1", "2019-11-11", '2019-11-11', "2019-11-10", 3, 1),
-    (9, "Spens event 2", "Opis spens eventa 2", "2019-9-10", "2019-9-11", "2019-10-5", 3, 1),
-    (10, "Spens event 3", "Opis spens eventa 3", "2019-8-10", "2019-8-11", "2019-10-5", 3, 1),
-    (11, "Spens event 4", "Opis spens eventa 4", "2019-7-10", "2019-7-11", "2019-10-5", 3, 1),
-    (12, "Spens event 5", "Opis spens eventa 5", "2019-6-10", "2019-6-11", "2019-10-5", 3, 1),		-- ova tri su istog meseca
-    (13, "Spens event 6", "Opis spens eventa 6", "2019-6-10", "2019-6-11", "2019-10-5", 3, 1),		--
-    (14, "Spens event 7", "Opis spens eventa 7", "2019-6-10", "2019-6-11", "2019-10-5", 3, 1);		--
+    (5, 'Spens event 1', 'Opis spens eventa 1', 'SPORTS', '2019-11-12', '2019-11-12', '2019-11-11', 3, 1),
+    (6, 'Spens event 2', 'Opis spens eventa 2', 'SPORTS', '2019-09-10', '2019-09-11', '2019-09-5', 3, 1),
+    (7, 'Spens event 3', 'Opis spens eventa 3', 'SPORTS', '2019-08-10', '2019-08-11', '2019-08-5', 3, 1),
+    (8, 'Spens event 4', 'Opis spens eventa 4', 'CULTURAL', '2019-07-10', '2019-07-11', '2019-07-5', 3, 1),
+    (9, 'Spens event 5', 'Opis spens eventa 5', 'CULTURAL', '2019-06-10', '2019-06-11', '2019-06-5', 3, 1),
+    (10, 'Spens event 6', 'Opis spens eventa 6', 'ENTERTAINMENT', '2019-06-12', '2019-06-13', '2019-06-5', 3, 1),
+    (11, 'Spens event 7', 'Opis spens eventa 7', 'ENTERTAINMENT', '2019-06-14', '2019-06-15', '2019-06-5', 3, 1);
 
 INSERT INTO users
     (dtype, id, email, first_name, last_name, password, username, is_verified, verification_code)
 VALUES
-    ('REGISTERED_USER', 23, 'kacjica+1@gmail.com', 'Katarina', 'Tukelic', 'kaca', 'kaca', 1, null);
+    ('REGISTERED_USER', 1, 'kacjica+1@gmail.com', 'Katarina', 'Tukelic', 'kaca', 'kaca', 1, null);
 
 INSERT INTO users
     (dtype, id, email, first_name, last_name, password, username)
 VALUES
-    ('ADMIN', 24, 'f.ivkovic16+1@gmail.com', 'Filip', 'Ivkovic', 'filip', 'filip');
+    ('ADMIN', 2, 'f.ivkovic16+1@gmail.com', 'Filip', 'Ivkovic', 'filip', 'filip');
 
 INSERT INTO user_role
 VALUES
-    (23, 'REGISTERED_USER'),
-    (24, 'ADMIN');
+    (1, 'REGISTERED_USER'),
+    (2, 'ADMIN');
 
-INSERT INTO ticketist.tickets
+INSERT INTO tickets
 	(id, is_paid, number_row, number_column, price, event_id, user_id)
 VALUES
-	(15, true, 2, 2, 20, 4, 23),
-    (16, true, 2, 3, 20, 4, 23),
-    (17, true, 2, 4, 20, 4, 23),
-    (18, false, 5, 5, 50, 5, 23),	-- ako je is_paid false, radi se o rezervaciji
+	(1, true, 2, 2, 20, 1, 1),
+    (2, true, 2, 3, 20, 1, 1),
+    (3, true, 2, 4, 20, 1, 1),
+    (4, false, 5, 5, 50, 2, 1),	  -- rezervacija
 
-    (19, false, 2, 2, 20, 8, 23),
-    (20, true, 2, 3, 20, 9, 23),
-    (21, true, 2, 4, 20, 12, 23),
-    (22, true, 2, 5, 20, 13, 23);
+    (5, false, 2, 2, 20, 5, 1),   -- rezervacija
+    (6, true, 2, 3, 20, 6, 1),
+    (7, true, 2, 4, 20, 9, 1),
+    (8, true, 2, 5, 20, 10, 1);
